dash-sample/requirements.txt
dash
pandas
python-dotenv
sqlalchemy
pymysql